refactor(frontend): move review component to separate module
